Fix ngspice download URL

+++ PySpice/Scripts/pyspice_post_installation.py
@@ -109,7 +109,9 @@ class PySpicePostInstallation:
     NGSPICE_BASE_URL = 'https://sourceforge.net/projects/ngspice/files'
     NGSPICE_RELEASE_URL = NGSPICE_BASE_URL + '/ng-spice-rework'
     NGSPICE_WINDOWS_DLL_URL = NGSPICE_RELEASE_URL + '/{0}/ngspice-{0}_dll_64.zip'
+    NGSPICE_WINDOWS_DLL_OLD_URL = NGSPICE_RELEASE_URL + '/old-releases/{0}/ngspice-{0}_dll_64.zip'
     NGSPICE_MANUAL_URL = NGSPICE_RELEASE_URL + '/{0}/ngspice-{0}-manual.pdf/download'
+    NGSPICE_MANUAL_OLD_URL = NGSPICE_RELEASE_URL + '/old-releases/{0}/ngspice-{0}-manual.pdf/download'
 
     ##############################################
 
@@ -177,10 +179,11 @@ class PySpicePostInstallation:
 
     ##############################################
 
-    def _donwload_file(self, url, dst_path):
+    def _download_file(self, url, dst_path):
         print('Get {} ... -> {}'.format(url, dst_path))
         response = requests.get(url, allow_redirects=True)
-        assert(response.status_code == requests.codes.ok)
+        if response.status_code != requests.codes.ok:
+            response.raise_for_status()
         with open(dst_path, mode='wb') as fh:
             fh.write(response.content)
 
@@ -210,7 +213,12 @@ class PySpicePostInstallation:
             url = self.NGSPICE_WINDOWS_DLL_URL.format(self.ngspice_version)
             zip_path = tmp_directory.joinpath('ngspice-{}_dll_64.zip'.format(self.ngspice_version))
             dst_path = Path(NgSpice.__file__).parent
-            self._donwload_file(url, zip_path)
+            try:
+                self._download_file(url, zip_path)
+            except requests.exceptions.HTTPError:
+                print('Download failed, trying another URL...')
+                url = self.NGSPICE_WINDOWS_DLL_OLD_URL.format(self.ngspice_version)
+                self._download_file(url, zip_path)
             with ZipFile(zip_path) as zip_file:
                 zip_file.extractall(path=dst_path)
                 print('Extracted {} in {}'.format(zip_path, dst_path.joinpath('Spice64_dll')))
@@ -220,17 +228,19 @@ class PySpicePostInstallation:
         # src = dll_path.joinpath('ngspice-{}.dll'.format(self.ngspice_version))
         src = 'ngspice-{}.dll'.format(self.ngspice_version)
         target = dll_path.joinpath('ngspice.dll')
-        if target.exists():
-            target.unlink()
-        try:
-            target.symlink_to(src)
-        except OSError:
-            # OSError: symbolic link privilege not held
-            # Windows: If User Account Control (UAC) is on, any user with the "Create Symbolic
-            #   Links" privilege that is not in the Administrators group can simply create a
-            #   symbolic link.  For users within the Administrators group and with UAC on, the user
-            #   must "Run as Administrator".
-            shutil.copy(target.parent.joinpath(src), target)
+        # For ngspice version <=31 DLL naming did not contain a version number
+        if dll_path.joinpath(src).exists():
+            if target.exists():
+                target.unlink()
+            try:
+                target.symlink_to(src)
+            except OSError:
+                # OSError: symbolic link privilege not held
+                # Windows: If User Account Control (UAC) is on, any user with the "Create Symbolic
+                #   Links" privilege that is not in the Administrators group can simply create a
+                #   symbolic link.  For users within the Administrators group and with UAC on, the user
+                #   must "Run as Administrator".
+                shutil.copy(target.parent.joinpath(src), target)
 
         spinit_path = spice64_path.joinpath('share', 'ngspice', 'scripts', 'spinit')
         with open(spinit_path) as fh:
@@ -251,7 +261,12 @@ class PySpicePostInstallation:
 
     def download_ngspice_manual(self):
         url = self.NGSPICE_MANUAL_URL.format(self.ngspice_version)
-        self._donwload_file(url, 'ngspice-manual-{}.pdf'.format(self.ngspice_version))
+        try:
+            self._download_file(url, 'ngspice-manual-{}.pdf'.format(self.ngspice_version))
+        except requests.exceptions.HTTPError:
+            print('Download failed, trying another URL...')
+            url = self.NGSPICE_MANUAL_OLD_URL.format(self.ngspice_version)
+            self._download_file(url, 'ngspice-manual-{}.pdf'.format(self.ngspice_version))
 
     ##############################################
 
@@ -438,7 +453,7 @@ class PySpicePostInstallation:
             return
 
         with tempfile.TemporaryDirectory() as tmp_directory:
-            self._donwload_file(RELEASE_URL, zip_path)
+            self._download_file(RELEASE_URL, zip_path)
             with ZipFile(zip_path) as zip_file:
                 zip_file.extractall(path=tmp_directory)
             examples_path = Path(tmp_directory).joinpath('PySpice-{}'.format(version), 'examples')
